+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* xx.xx.xx
+  - fix bug in IndicatorBox proximal_conjugate
 * 23.0.1
   - Fix bug with NikonReader requiring ROI to be set in constructor.